Ui for user is created. Navbar changed
UI elements for users are ready. I also changed the navbar names. I did not change the component name in vue code because it was giving me errors that I could care less about. so keep that in mind, components have weird names but who gives a fuck.
credit: chatgpt